+++ psychoshare_api/DTOs/Ban/BanResponseDto.cs
@@ -1,7 +1,7 @@
 public class BanResponseDto
 {
-    public int IdBan { get; set; }
-    public int IdPerson { get; set; }
+    public long IdBan { get; set; }
+    public long IdPerson { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
     public string Reason { get; set; } = "";
